pyproject.toml
[project]
name = "percolation-hypotheses-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.9",
    "click>=8.2.0",
    "google-adk>=0.5.0",
    "google-genai>=1.15.0",
    "langchain-text-splitters>=0.3.8",
    "nltk>=3.9.1",
    "numpy>=2.2.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "spacy>=3.8.7",
    "streamlit>=1.45.1",
    "textstat>=0.7.7",
]
